VTK: make sophisticated algorithm for enumerating border cells
Idea is that big mesh get split by 2 in all dimensions. Then if leaf is
smaller than given threshold or it contains only boundary it is
processed directly (by iterating over all cells). Otherwise a test for
existence of masked and non-masked cells is executed over the leaf. If
leaf contains both types of cells it undergo further split. If it
contains only masked or only non-masked cells then only boundaries of
the leaf are processed.
The minimum leaf size that is possible to split can be set via new
parameter that is exported to Python and can be controlled from client
side.